Converted create_community to use social links field
