+++ packages/commonwealth/client/scripts/views/pages/manage_community/community_metadata_rows.tsx
@@ -14,15 +14,13 @@ import type CommunityInfo from '../../../models/ChainInfo';
 import type RoleInfo from '../../../models/RoleInfo';
 import { AvatarUpload } from '../../components/Avatar';
 import { CWButton } from '../../components/component_kit/cw_button';
-import { CWDropdown } from '../../components/component_kit/cw_dropdown';
 import { CWIcon } from '../../components/component_kit/cw_icons/cw_icon';
-import { CWClose } from '../../components/component_kit/cw_icons/cw_icons';
 import { CWLabel } from '../../components/component_kit/cw_label';
 import { CWSpinner } from '../../components/component_kit/cw_spinner';
 import { CWText } from '../../components/component_kit/cw_text';
 import { CWToggle } from '../../components/component_kit/cw_toggle';
-import { openConfirmation } from '../../modals/confirmation_modal';
 import DirectoryPageSection from './DirectoryPageSection';
+import { DiscordForumConnections } from './DiscordForumConnections';
 import { setCommunityCategories, setSelectedTags } from './helpers';
 import { ManageRoles } from './manage_roles';
 
@@ -32,123 +30,6 @@ type CommunityMetadataRowsProps = {
   mods: any;
   onRoleUpdate: (oldRole: RoleInfo, newRole: RoleInfo) => void;
   onSave: () => void;
-};
-
-type DiscordChannelConnection = {
-  channelName: string;
-  channelId: string;
-  onConnect: (topicId: string) => void;
-};
-
-const DiscordForumConnections = ({
-  channels,
-  topics,
-  refetchTopics,
-}: {
-  channels: DiscordChannelConnection[];
-  topics: { id: string; name: string; channelId: string | null }[];
-  refetchTopics: () => Promise<void>;
-}) => {
-  const topicOptions = topics.map((topic) => {
-    return { label: topic.name, value: topic.id };
-  });
-
-  const connectedTopics = topics.filter(
-    (topic) => topic.channelId !== null && topic.channelId !== '',
-  );
-
-  const [connectionVerified, setConnectionVerified] = useState(true);
-
-  const removeConnection = (topicId: string) => {
-    openConfirmation({
-      title: 'Warning',
-      // eslint-disable-next-line max-len
-      description: `Are you sure you want to remove this connection? New comments on Discord threads will NOT be updated on Commonwealth.`,
-      buttons: [
-        {
-          label: 'Remove',
-          buttonType: 'destructive',
-          buttonHeight: 'sm',
-          onClick: async () => {
-            try {
-              await app.discord.setForumChannelConnection(topicId, null);
-              setConnectionVerified(false);
-              await refetchTopics();
-              setConnectionVerified(true);
-            } catch (e) {
-              console.log(e);
-            }
-          },
-        },
-        {
-          label: 'No',
-          buttonType: 'secondary',
-          buttonHeight: 'sm',
-        },
-      ],
-    });
-  };
-
-  return (
-    <div className="DiscordForumConnections">
-      {channels.length > 0 ? (
-        <div className="TopicRow">
-          <CWText className="HeaderText">Channel</CWText>
-          <CWText>Topic</CWText>
-        </div>
-      ) : (
-        <CWText>Add a forum channel in your Discord server</CWText>
-      )}
-
-      {channels.map((channel) => {
-        const connectedTopic = topics.find(
-          (topic) => topic.channelId === channel.channelId,
-        );
-
-        const remainingTopics = topicOptions.filter(
-          (topic) =>
-            !connectedTopics.find(
-              (connected_topic) => connected_topic.id === topic.value,
-            ),
-        );
-
-        if (connectedTopic) {
-          remainingTopics.push({
-            label: connectedTopic.name,
-            value: connectedTopic.id,
-          });
-        }
-
-        return (
-          <div key={channel.channelId} className="TopicRow">
-            <CWText className="ChannelText">#{channel.channelName}</CWText>
-            {connectionVerified && (
-              <CWDropdown
-                initialValue={
-                  connectedTopic
-                    ? { label: connectedTopic.name, value: connectedTopic.id }
-                    : { label: 'Not connected', value: '' }
-                }
-                options={remainingTopics}
-                onSelect={async (item) => {
-                  // Connect the channel to the topic
-                  channel.onConnect(item.value);
-                }}
-              />
-            )}
-            {connectedTopic && (
-              <CWClose
-                className="CloseButton"
-                onClick={() => {
-                  removeConnection(connectedTopic.id);
-                }}
-              />
-            )}
-          </div>
-        );
-      })}
-    </div>
-  );
 };
 
 export const CommunityMetadataRows = ({
@@ -223,12 +104,6 @@ export const CommunityMetadataRows = ({
     community.communityBanner,
   );
   const [uploadInProgress, setUploadInProgress] = useState(false);
-  const [snapshotNotificationsEnabled, setSnapshotNotificationsEnabled] =
-    useState(false);
-  const [selectedSnapshotChannel, setSelectedSnapshotChannel] = useState<{
-    id: string;
-    name: string;
-  } | null>(null);
   const [discordWebhooksEnabled, setDiscordWebhooksEnabled] = useState(
     community.discordBotWebhooksEnabled,
   );
@@ -236,17 +111,6 @@ export const CommunityMetadataRows = ({
   useEffect(() => {
     setDiscordBotConnected(community.discordConfigId !== null);
   }, [community]);
-
-  useEffect(() => {
-    if (
-      discordChannels &&
-      discordChannels.selectedChannel &&
-      discordChannels.selectedChannel.id
-    ) {
-      setSelectedSnapshotChannel(discordChannels.selectedChannel);
-      setSnapshotNotificationsEnabled(true);
-    }
-  }, [discordChannels]);
 
   if (!selectedTags2) {
     return;
@@ -650,38 +514,6 @@ export const CommunityMetadataRows = ({
               className="connect-button"
               onClick={handleReconnectBot}
             />
-            {/* <div className="snapshot-settings">
-              <CWText type="h4">Snapshot Notifications</CWText>
-              <CWToggle
-                onChange={() =>
-                  setSnapshotNotificationsEnabled((prevState) => !prevState)
-                }
-                checked={snapshotNotificationsEnabled}
-              />
-            </div>
-            <div className="connected-line">
-              {discordChannels && snapshotNotificationsEnabled && (
-                <CWDropdown
-                  label={'Select Channel'}
-                  options={discordChannels.textChannels.map((channel) => {
-                    return {
-                      label: channel.name,
-                      value: channel.id,
-                    };
-                  })}
-                  initialValue={{
-                    value: selectedSnapshotChannel?.id ?? 'channel',
-                    label: selectedSnapshotChannel?.name ?? 'Select a Channel',
-                  }}
-                  onSelect={(item) => {
-                    setSelectedSnapshotChannel({
-                      id: item.value,
-                      name: item.label,
-                    });
-                  }}
-                />
-              )}
-            </div> */}
             <div className="snapshot-settings">
               <CWText type="h4">Connected Forum Channels</CWText>
             </div>
