+++ packages/commonwealth/test/integration/api/updateChain.spec.ts
@@ -16,7 +16,8 @@ const baseRequest: CommunityAttributes = {
   base: ChainBase.Substrate,
   icon_url: '/static/img/protocols/edg.png',
   active: true,
-  type: ChainType.Chain
+  type: ChainType.Chain,
+  social_links: [],
 };
 
 describe('UpdateChain Tests', () => {
@@ -26,29 +27,27 @@ describe('UpdateChain Tests', () => {
 
   it('Correctly updates chain', async () => {
     const controller = new ServerCommunitiesController(models, null, null);
-    const user: UserInstance =
-      buildUser({ models, userAttributes: { email: '', id: 1, isAdmin: true } }) as UserInstance;
+    const user: UserInstance = buildUser({
+      models,
+      userAttributes: { email: '', id: 1, isAdmin: true },
+    }) as UserInstance;
 
-    let response = await controller.updateCommunity(
-      {
-        directory_page_enabled: true,
-        directory_page_chain_node_id: 1,
-        ...baseRequest,
-        user: user
-      },
-    );
+    let response = await controller.updateCommunity({
+      directory_page_enabled: true,
+      directory_page_chain_node_id: 1,
+      ...baseRequest,
+      user: user,
+    });
 
     assert.equal(response.directory_page_enabled, true);
     assert.equal(response.directory_page_chain_node_id, 1);
 
-    response = await controller.updateCommunity(
-      {
-        directory_page_enabled: false,
-        directory_page_chain_node_id: null,
-        ...baseRequest,
-        user: user
-      },
-    );
+    response = await controller.updateCommunity({
+      directory_page_enabled: false,
+      directory_page_chain_node_id: null,
+      ...baseRequest,
+      user: user,
+    });
 
     assert.equal(response.directory_page_enabled, false);
     assert.equal(response.directory_page_chain_node_id, null);
