+++ packages/commonwealth/test/integration/api/updatecommunity.spec.ts
@@ -95,24 +95,14 @@ describe('Update Community/Chain Tests', () => {
       expect(res.body.result.discord).to.be.equal(discord);
     });
 
-    it.skip('should fail to update github without proper prefix', async () => {
-      const github = 'github.com';
+    it.skip('should fail to update social link without proper prefix', async () => {
+      const socialLinks = ['github.com'];
       const res = await chai
         .request(app)
         .patch(`/api/communities/${chain}`)
         .set('Accept', 'application/json')
-        .send({ jwt: jwtToken, id: chain, github });
-      expect(res.body.error).to.be.equal(ChainError.InvalidGithub);
-    });
-
-    it.skip('should fail to update telegram without proper prefix', async () => {
-      const telegram = 't.me';
-      const res = await chai
-        .request(app)
-        .patch(`/api/communities/${chain}`)
-        .set('Accept', 'application/json')
-        .send({ jwt: jwtToken, id: chain, telegram });
-      expect(res.body.error).to.be.equal(ChainError.InvalidTelegram);
+        .send({ jwt: jwtToken, id: chain, socialLinks });
+      expect(res.body.error).to.exist;
     });
 
     it('should update telegram', async () => {
